fix(server): Fix linter warnings and formatting issues
Resolves all Go linter warnings and Prettier formatting issues:

- Use Go 1.22+ integer range syntax in concurrency tests
- Replace if-else chains with switch statements for better readability
- Convert struct types instead of using struct literals
- Remove unnecessary type conversions (EdgeHandle is already int32)
- Add safe int32→uint64 conversion with overflow protection
- Remove unused publishBulkEdgeDelete and publishBulkFlowVariableDelete functions
- Format markdown files with Prettier

All linter checks now pass with 0 issues.

+++ packages/server/internal/api/rflowv2/rflowv2_edge.go
@@ -200,7 +200,7 @@ func (s *FlowServiceV2RPC) EdgeUpdate(ctx context.Context, req *connect.Request[
 
 		if item.SourceHandle != nil {
 			existing.SourceHandler = convertHandle(item.GetSourceHandle())
-			edgePatch.SourceHandler = patch.NewOptional(int32(existing.SourceHandler))
+			edgePatch.SourceHandler = patch.NewOptional(existing.SourceHandler)
 		}
 
 		validatedUpdates = append(validatedUpdates, updateData{
